+++ Tests/OverlayScrollViewDelegateTest.swift
@@ -16,15 +16,27 @@ private class ScrollViewController: UIViewController, UIScrollViewDelegate {
 
 class OverlayScrollViewDelegateTest: QuickSpec {
     override func spec() {
-        let viewController = ScrollViewController()
-        let overlayContainer = OverlayContainerViewController()
-        overlayContainer.viewControllers = [viewController]
-        overlayContainer.loadViewIfNeeded()
+        var viewController: ScrollViewController!
+        var overlayContainer: OverlayContainerViewController!
+        beforeEach {
+            viewController = ScrollViewController()
+            overlayContainer = OverlayContainerViewController()
+            overlayContainer.viewControllers = [viewController]
+            overlayContainer.loadViewIfNeeded()
+        }
         it("should observe the scroll view delegate") {
             overlayContainer.drivingScrollView = viewController.scrollView
             expect(viewController.scrollView.delegate).toNot(beNil())
             viewController.scrollView.delegate = viewController
             expect(viewController.scrollView.delegate).toNot(beIdenticalTo(viewController))
+        }
+        it("should restore the scroll view delegate") {
+            viewController.scrollView.delegate = viewController
+            expect(viewController.scrollView.delegate).to(beIdenticalTo(viewController))
+            overlayContainer.drivingScrollView = viewController.scrollView
+            expect(viewController.scrollView.delegate).toNot(beIdenticalTo(viewController))
+            overlayContainer.drivingScrollView = nil
+            expect(viewController.scrollView.delegate).to(beIdenticalTo(viewController))
         }
     }
 }